requirements.txt
# Recomenda-se usar um ambiente virtual (venv ou conda) para isolar as dependências do projeto.
pip>=26.0.0

# Dependências principais para manipulação de dados, modelagem e serialização
pandas>=2.2.0 # Updated to 2.2.0 as it's common and compatible
numpy>=1.26.0 # Updated to 1.26.0 as it's common and compatible
scikit-learn>=1.4.0 # Updated to a more stable, current version for better compatibility
joblib>=1.3

# Visualização (se você rodar seu main.py ou fizer análises exploratórias)
matplotlib>=3.8

# MLflow para rastreamento de experimentos, gerenciamento de modelos e deployment
mlflow>=2.11.0 # Updated to a more stable, current version
PyYAML>=6.0

# (Opcional) Dependências para deployment com FastAPI e monitoramento com Prometheus
fastapi>=0.110
uvicorn[standard]>=0.28
prometheus-client>=0.19
pydantic>=2.6
pydantic-settings>=2.2
psycopg2-binary>=2.9
python-multipart>=0.0.6

# Dependências para a aplicação Streamlit
streamlit>=1.30.0

# Dependências para integração com AWS (se for usar S3 para armazenamento de artefatos ou SageMaker para deployment)    
boto3>=1.34
botocore>=1.34

# Dependências para testes unitários e de integração
pytest>=8.0
iniconfig>=2.0
cloudpickle>=3.0
